Player Base, Command System, and Network Base Added
Basic player class created, just holds the data needed for the player to function in combat at this point. File I/O for reading in the list of commands each job can use, network base set up for when that needs to happen. Next steps on my end are fleshing out reading in input from the board to check against the command list, and then doing actions against that, which will finish up combat on the player end, and then sort out how were gonna have our second player in the end.